refactor(types): вынести KeywordMetadataLoader в отдельный класс (Sonar)
В одном раунде закрываю 11 Sonar-issues, появившихся после первого захода с
keywords-фолбэком:

  S2972 — KeywordMetadataBuilder был inner-классом > 25 строк: вынесен
    наружу как top-level package-private KeywordMetadataLoader + Builder.
    Записи KeywordMetadata тоже вынесена в отдельный файл.
  S3398 ×3 — readSnippet/readKeywordDescription/readDescriptionByParent
    логически принадлежали Builder'у — переехали туда же как private static
    методы.
  S1213 ×3 — мои локальные константы JSON_FIELD_* убраны из GSP; в
    KeywordMetadataLoader константы лежат блоком в начале класса, в нужном
    порядке.
  S1905 — лишний явный cast (Map<String,Object>) убран — вместо этого
    типизированный helper readRoot, единственный @SuppressWarnings.
  S1309 ×2 — два оставшихся @SuppressWarnings: один на readRoot
    (cast на raw Map.class из Jackson — без него никак), второй на
    readDescriptionByParent (на проход по nested Map<String,Map<String,String>>).
    Локализованы на конкретных строках, а не растянуты по методу.
  S1166 — IOException теперь логируется аргументом (стек попадает в лог).

В GSP остаётся только тонкая обёртка loadKeywordMetadata, делегирующая в
KeywordMetadataLoader.load. Конечный API не изменился — Loaded record
по-прежнему получает keywords/scopes/snippets/descriptions через
keywordMeta.X() (record accessors), тесты регистрации зелёные.
