+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/types/registry/GlobalScopeProvider.java
@@ -74,6 +74,7 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Function;
+import java.util.function.Predicate;
 import java.util.function.Supplier;
 
 /**
@@ -1169,16 +1170,16 @@ public class GlobalScopeProvider {
       var keywordMeta = loadKeywordMetadata(keywordsResourceFor(resourcePath), scope);
       var seenKeywords = new HashSet<String>();
       keywords.forEach(k -> seenKeywords.add(k.toLowerCase(Locale.ROOT)));
-      for (var k : keywordMeta.keywords) {
+      for (var k : keywordMeta.keywords()) {
         if (seenKeywords.add(k.toLowerCase(Locale.ROOT))) {
           keywords.add(k);
         }
       }
-      kwScopes.putAll(keywordMeta.scopes);
+      kwScopes.putAll(keywordMeta.scopes());
       return new Loaded(functions, List.copyOf(classes), List.copyOf(keywords), variables,
         List.of(),
         fnScopes, clsScopes, kwScopes, varScopes,
-        Map.copyOf(keywordMeta.snippets), Map.copyOf(keywordMeta.descriptions));
+        Map.copyOf(keywordMeta.snippets()), Map.copyOf(keywordMeta.descriptions()));
     } catch (IOException e) {
       LOGGER.error("Failed to load builtin globals resource: {}", resourcePath, e);
       return new Loaded(Collections.emptyMap(), List.of(), List.of(), List.of(),
@@ -1195,172 +1196,11 @@ public class GlobalScopeProvider {
     return KEYWORDS_RESOURCE_PATH;
   }
 
-  /** Результат чтения builtin-keywords.json — keywords + scopes + snippets + descriptions. */
-  private record KeywordMetadata(List<String> keywords, Map<String, LanguageScope> scopes,
-                                 Map<String, LanguageKeywordSnippet> snippets,
-                                 Map<String, KeywordDescription> descriptions) {
-    static final KeywordMetadata EMPTY = new KeywordMetadata(List.of(), Map.of(), Map.of(), Map.of());
-  }
+  private static final Predicate<String> INCLUDE_IN_COMPLETION =
+    categoryStr -> KEYWORD_CATEGORIES.stream().anyMatch(c -> c.name().equals(categoryStr));
 
-  private static final String JSON_FIELD_DESCRIPTION = "description";
-  private static final String JSON_FIELD_RU = "ru";
-  private static final String JSON_FIELD_EN = "en";
-
-  /**
-   * Парсит JSON-fallback keywords. Поле {@code keywords} — список объектов
-   * с обязательным {@code name} и опциональными {@code alias}, {@code category},
-   * {@code description}, {@code descriptionEn}, {@code snippet.ru/en},
-   * {@code descriptionByParent.<parent>.{ru,en}}.
-   * <p>
-   * Фильтр совпадает с {@link #KEYWORD_CATEGORIES} из bsl-context (no-dot
-   * completion для LITERAL/STATEMENT/OPERATOR/DECLARATION); PRAGMA / ANNOTATION /
-   * PREPROCESSOR_INSTRUCTION в плоский список keywords не попадают, но
-   * описание/сниппет для них всё равно индексируются (используются hover'ом
-   * в соответствующих синтаксических контекстах).
-   */
-  @SuppressWarnings("unchecked")
   private static KeywordMetadata loadKeywordMetadata(String resourcePath, LanguageScope scope) {
-    var mapper = JsonMapper.builder().build();
-    try (var stream = new ClassPathResource(resourcePath).getInputStream()) {
-      var root = (Map<String, Object>) mapper.readValue(stream, Map.class);
-      var raw = (List<Map<String, Object>>) root.getOrDefault("keywords", Collections.emptyList());
-      var builder = new KeywordMetadataBuilder(scope);
-      for (var entry : raw) {
-        builder.add(entry);
-      }
-      return builder.build();
-    } catch (IOException e) {
-      LOGGER.warn("Builtin keywords resource not found or unreadable: {}", resourcePath, e);
-      return KeywordMetadata.EMPTY;
-    }
-  }
-
-  /**
-   * Аккумулирует разбор {@code builtin-keywords.json} в {@link KeywordMetadata}:
-   * плоский список имён для completion, scopes по локали, snippet'ы и
-   * описания (включая {@code descriptionByParent}).
-   */
-  private static final class KeywordMetadataBuilder {
-    private final LanguageScope scope;
-    private final Set<String> seen = new HashSet<>();
-    private final List<String> keywords = new ArrayList<>();
-    private final Map<String, LanguageScope> scopes = new HashMap<>();
-    private final Map<String, LanguageKeywordSnippet> snippets = new HashMap<>();
-    private final Map<String, KeywordDescription> descriptions = new HashMap<>();
-
-    KeywordMetadataBuilder(LanguageScope scope) {
-      this.scope = scope;
-    }
-
-    void add(Map<String, Object> entry) {
-      var name = stringField(entry, "name");
-      if (name.isBlank()) {
-        return;
-      }
-      var alias = stringField(entry, "alias");
-      if (isCompletionCategory(stringField(entry, "category"))) {
-        register(name);
-        if (!alias.isBlank()) {
-          register(alias);
-        }
-      }
-      indexSnippet(entry, name, alias);
-      indexDescription(entry, name, alias);
-    }
-
-    KeywordMetadata build() {
-      return new KeywordMetadata(List.copyOf(keywords), Map.copyOf(scopes),
-        Map.copyOf(snippets), Map.copyOf(descriptions));
-    }
-
-    private void register(String written) {
-      if (seen.add(written.toLowerCase(Locale.ROOT))) {
-        keywords.add(written);
-        scopes.put(written.toLowerCase(Locale.ROOT), scope);
-      }
-    }
-
-    private void indexSnippet(Map<String, Object> entry, String name, String alias) {
-      var snippet = readSnippet(entry);
-      if (snippet.isEmpty()) {
-        return;
-      }
-      snippets.put(name.toLowerCase(Locale.ROOT), snippet);
-      if (!alias.isBlank()) {
-        snippets.put(alias.toLowerCase(Locale.ROOT), snippet);
-      }
-    }
-
-    private void indexDescription(Map<String, Object> entry, String name, String alias) {
-      var description = readKeywordDescription(entry);
-      if (description.isEmpty()) {
-        return;
-      }
-      descriptions.put(name.toLowerCase(Locale.ROOT), description);
-      if (!alias.isBlank()) {
-        descriptions.put(alias.toLowerCase(Locale.ROOT), description);
-      }
-    }
-  }
-
-  private static boolean isCompletionCategory(String categoryStr) {
-    return KEYWORD_CATEGORIES.stream().anyMatch(c -> c.name().equals(categoryStr));
-  }
-
-  private static LanguageKeywordSnippet readSnippet(Map<String, Object> entry) {
-    var raw = asStringMap(entry.get("snippet"));
-    if (raw == null) {
-      return LanguageKeywordSnippet.EMPTY;
-    }
-    var ru = raw.getOrDefault(JSON_FIELD_RU, "");
-    var en = raw.getOrDefault(JSON_FIELD_EN, "");
-    if (ru.isEmpty() && en.isEmpty()) {
-      return LanguageKeywordSnippet.EMPTY;
-    }
-    return new LanguageKeywordSnippet(ru, en);
-  }
-
-  private static KeywordDescription readKeywordDescription(Map<String, Object> entry) {
-    var primary = BilingualString.of(
-      stringField(entry, JSON_FIELD_DESCRIPTION),
-      stringField(entry, "descriptionEn"));
-    var byParent = readDescriptionByParent(entry.get("descriptionByParent"));
-    if (primary.isEmpty() && byParent.isEmpty()) {
-      return KeywordDescription.EMPTY;
-    }
-    return new KeywordDescription(primary, byParent);
-  }
-
-  @SuppressWarnings("unchecked")
-  private static Map<String, BilingualString> readDescriptionByParent(@Nullable Object raw) {
-    if (!(raw instanceof Map<?, ?> rawMap) || rawMap.isEmpty()) {
-      return Map.of();
-    }
-    var result = new LinkedHashMap<String, BilingualString>();
-    for (var e : ((Map<String, Map<String, String>>) raw).entrySet()) {
-      var pair = e.getValue();
-      if (pair == null) {
-        continue;
-      }
-      var bi = BilingualString.of(
-        pair.getOrDefault(JSON_FIELD_RU, ""),
-        pair.getOrDefault(JSON_FIELD_EN, ""));
-      if (!bi.isEmpty()) {
-        result.put(e.getKey(), bi);
-      }
-    }
-    return result;
-  }
-
-  private static String stringField(Map<String, Object> entry, String key) {
-    var value = entry.get(key);
-    return value instanceof String s ? s : "";
-  }
-
-  @SuppressWarnings("unchecked")
-  @Nullable
-  private static Map<String, String> asStringMap(@Nullable Object raw) {
-    return raw instanceof Map<?, ?> ? (Map<String, String>) raw : null;
+    return KeywordMetadataLoader.load(resourcePath, scope, INCLUDE_IN_COMPLETION);
   }
 
   @SuppressWarnings("unchecked")
